RC1 - adding blacklist of authors
Bookport started to offer plenty of books by the same authors, in various translations, that the libary may not like. A (black)list of authors has been added. Set them in variable: authors2del (bash array) at the begging of the script. If this name is found in Marc 100 field, the whole record is deleted from import dataset.